requirements.txt
fpdf2==2.5.6
pyttsx3==2.90
requests==2.31.0
python-dotenv==1.0.0
mistralai==0.1.7
